Add support for subsequent guesses
Now the bot can be used to help humans through a whole game.

Store guesses in hash set instead of vec to efficiently check if a
guess is also an answer.

Rewrite some functions to take a string slice instead of char because
I am still a bit unsure how hashing works in rust and I initially
suspected this to be the reason why the program always said that the
no guesses were also answers. This turned out to be a problem with the
wordlists and I have not reverted the code. If memory serves the
program was faster before this change, if true this seems a likely
culprit.