Add allowed answers to guesses
Because some games list the guesses without the answers, presumably to
keep the redundancy down, and not being allowed to guess on a potential
answer is absurd.

Change the word listed as optimal for absurdle because with the updated
list of guesses the previous word is no longer optimal. Also update the
wording to reflect that the optimal word is usually not unique.