Fix cargo publish thinking repo is dirty
I have previously experienced a similar problem where touching a file
without updating its content caused `git diff` to exit with an error.
In that case doing `git update-index --refresh` solved the issue. At
this point I cannot remember how this worked but it is not helping.
However, when troubleshooting I found that `git status` made the
problem go away so I will stick with that for now.