Adapt program for wapm
Bundle the most popular wordlist because interacting with the file-
system is less straight forward in wasm, especially when run in the
browser.

Remove rayon because it causes a panic.

Remove main return because with it the entrypoint is not recognized and
trying to run the main function explicitly reveals that this has an
interface that expects two integers. Haven't bothered looking into why
this is.